✨ feat(mes): 增加车间编码和主单位校验逻辑
新增车间编码常量，优化获取车间信息的逻辑，确保在删除计量单位时校验是否存在辅单位，提升数据一致性和安全性。

+++ yudao-module-mes/src/main/java/cn/iocoder/yudao/module/mes/service/md/unitmeasure/MesMdUnitMeasureServiceImpl.java
@@ -62,10 +62,15 @@ public class MesMdUnitMeasureServiceImpl implements MesMdUnitMeasureService {
     public void deleteUnitMeasure(Long id) {
         // 校验存在
         validateUnitMeasureExists(id);
+        // 校验是否存在以此为主单位的辅单位
+        if (unitMeasureMapper.selectCountByPrimaryId(id) > 0) {
+            throw exception(MD_UNIT_MEASURE_HAS_SECONDARY);
+        }
         // 校验是否被物料引用
         if (itemService.getItemCountByUnitMeasureId(id) > 0) {
             throw exception(MD_UNIT_MEASURE_HAS_ITEM);
         }
+
         // 删除
         unitMeasureMapper.deleteById(id);
     }
